+++ emote/mixins/__init__.py
@@ -1,0 +1,8 @@
+"""Mixins for emote.
+
+Mixins are used to add functionality to other classes just like
+regular inheritance. The difference is that mixins are designed to
+work well with multiple inheritance, which requires extra care to
+avoid issues in initialization order.
+
+"""