+++ emote/models/callbacks.py
@@ -6,9 +6,11 @@ import torch
 
 from torch import optim
 
-from emote.callbacks import BatchCallback, LossCallback
+from emote.callback import BatchCallback
+from emote.callbacks.loss import LossCallback
 from emote.extra.schedules import BPStepScheduler
 from emote.memory import MemoryLoader
+from emote.mixins.logging import LoggingMixin
 from emote.models.model import DynamicModel
 from emote.models.model_env import ModelEnv
 from emote.proxies import AgentProxy, MemoryProxy
@@ -60,7 +62,7 @@ class ModelLoss(LossCallback):
         return loss
 
 
-class LossProgressCheck(BatchCallback):
+class LossProgressCheck(LoggingMixin, BatchCallback):
     def __init__(
         self,
         model: DynamicModel,
@@ -196,7 +198,7 @@ class BatchSampler(BatchCallback):
         return True if rnd < self.prob_of_sampling_model_data else False
 
 
-class ModelBasedCollector(BatchCallback):
+class ModelBasedCollector(LoggingMixin, BatchCallback):
     """ModelBasedCollector class is used to sample rollouts from the trained dynamic model.
     The rollouts are stored in a replay buffer memory.
 
@@ -222,9 +224,9 @@ class ModelBasedCollector(BatchCallback):
         input_key: str = "obs",
     ):
         super().__init__()
-        """ The data group is used to receive correct observation when collect_multiple is 
-            called. The data group must be set such that real Gym samples (not model data) 
-            are given to the function. 
+        """ The data group is used to receive correct observation when collect_multiple is
+            called. The data group must be set such that real Gym samples (not model data)
+            are given to the function.
         """
         self.data_group = data_group
         self._input_key = input_key